tests: add utilities to parse known-to-be-valid IPs, CIDRs, and MACs
