+++ go-controller/hybrid-overlay/pkg/controller/master_test.go
@@ -2,7 +2,6 @@ package controller
 
 import (
 	"fmt"
-	"net"
 	"strings"
 
 	"github.com/urfave/cli"
@@ -22,14 +21,6 @@ import (
 )
 
 const hoNodeCliArg string = "-no-hostsubnet-nodes=" + v1.LabelOSStable + "=windows"
-
-func mustParseCIDR(cidr string) *net.IPNet {
-	_, net, err := net.ParseCIDR(cidr)
-	if err != nil {
-		panic("bad CIDR string constant " + cidr)
-	}
-	return net
-}
 
 func addGetPortAddressesCmds(fexec *ovntest.FakeExec, nodeName, hybMAC, hybIP string) {
 	addresses := hybMAC + " " + hybIP
