+++ go-controller/pkg/config/utils_test.go
@@ -3,12 +3,9 @@ package config
 import (
 	"net"
 	"testing"
-)
 
-func returnIPNetPointers(input string) *net.IPNet {
-	_, ipNet, _ := net.ParseCIDR(input)
-	return ipNet
-}
+	ovntest "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/testing"
+)
 
 func TestParseClusterSubnetEntries(t *testing.T) {
 	tests := []struct {
@@ -20,22 +17,22 @@ func TestParseClusterSubnetEntries(t *testing.T) {
 		{
 			name:            "Single CIDR correctly formatted",
 			cmdLineArg:      "10.132.0.0/26/28",
-			clusterNetworks: []CIDRNetworkEntry{{CIDR: returnIPNetPointers("10.132.0.0/26"), HostSubnetLength: 28}},
+			clusterNetworks: []CIDRNetworkEntry{{CIDR: ovntest.MustParseIPNet("10.132.0.0/26"), HostSubnetLength: 28}},
 			expectedErr:     false,
 		},
 		{
 			name:       "Two CIDRs correctly formatted",
 			cmdLineArg: "10.132.0.0/26/28,10.133.0.0/26/28",
 			clusterNetworks: []CIDRNetworkEntry{
-				{CIDR: returnIPNetPointers("10.132.0.0/26"), HostSubnetLength: 28},
-				{CIDR: returnIPNetPointers("10.133.0.0/26"), HostSubnetLength: 28},
+				{CIDR: ovntest.MustParseIPNet("10.132.0.0/26"), HostSubnetLength: 28},
+				{CIDR: ovntest.MustParseIPNet("10.133.0.0/26"), HostSubnetLength: 28},
 			},
 			expectedErr: false,
 		},
 		{
 			name:            "Test that defaulting to hostsubnetlength 8 still works",
 			cmdLineArg:      "10.128.0.0/14",
-			clusterNetworks: []CIDRNetworkEntry{{CIDR: returnIPNetPointers("10.128.0.0/14"), HostSubnetLength: 24}},
+			clusterNetworks: []CIDRNetworkEntry{{CIDR: ovntest.MustParseIPNet("10.128.0.0/14"), HostSubnetLength: 24}},
 			expectedErr:     false,
 		},
 		{
@@ -83,13 +80,13 @@ func TestParseClusterSubnetEntries(t *testing.T) {
 		{
 			name:            "IPv6",
 			cmdLineArg:      "fda6::/48/64",
-			clusterNetworks: []CIDRNetworkEntry{{CIDR: returnIPNetPointers("fda6::/48"), HostSubnetLength: 64}},
+			clusterNetworks: []CIDRNetworkEntry{{CIDR: ovntest.MustParseIPNet("fda6::/48"), HostSubnetLength: 64}},
 			expectedErr:     false,
 		},
 		{
 			name:            "IPv6 defaults to /64 hostsubnets",
 			cmdLineArg:      "fda6::/48",
-			clusterNetworks: []CIDRNetworkEntry{{CIDR: returnIPNetPointers("fda6::/48"), HostSubnetLength: 64}},
+			clusterNetworks: []CIDRNetworkEntry{{CIDR: ovntest.MustParseIPNet("fda6::/48"), HostSubnetLength: 64}},
 			expectedErr:     false,
 		},
 		{
@@ -147,58 +144,58 @@ func Test_checkForOverlap(t *testing.T) {
 		{
 			name: "non-overlapping",
 			cidrList: []*net.IPNet{
-				returnIPNetPointers("10.132.0.0/26"),
-				returnIPNetPointers("10.133.0.0/26"),
+				ovntest.MustParseIPNet("10.132.0.0/26"),
+				ovntest.MustParseIPNet("10.133.0.0/26"),
 			},
 			shouldError: false,
 		},
 		{
 			name: "duplicate entry",
 			cidrList: []*net.IPNet{
-				returnIPNetPointers("10.132.0.0/26"),
-				returnIPNetPointers("10.132.0.0/26"),
+				ovntest.MustParseIPNet("10.132.0.0/26"),
+				ovntest.MustParseIPNet("10.132.0.0/26"),
 			},
 			shouldError: true,
 		},
 		{
 			name: "proper subset",
 			cidrList: []*net.IPNet{
-				returnIPNetPointers("10.132.0.0/26"),
-				returnIPNetPointers("10.0.0.0/8"),
+				ovntest.MustParseIPNet("10.132.0.0/26"),
+				ovntest.MustParseIPNet("10.0.0.0/8"),
 			},
 			shouldError: true,
 		},
 		{
 			name: "proper superset",
 			cidrList: []*net.IPNet{
-				returnIPNetPointers("10.0.0.0/8"),
-				returnIPNetPointers("10.133.0.0/26"),
+				ovntest.MustParseIPNet("10.0.0.0/8"),
+				ovntest.MustParseIPNet("10.133.0.0/26"),
 			},
 			shouldError: true,
 		},
 		{
 			name: "overlap",
 			cidrList: []*net.IPNet{
-				returnIPNetPointers("10.1.0.0/14"),
-				returnIPNetPointers("10.0.0.0/15"),
+				ovntest.MustParseIPNet("10.1.0.0/14"),
+				ovntest.MustParseIPNet("10.0.0.0/15"),
 			},
 			shouldError: true,
 		},
 		{
 			name: "non-overlapping V4 ipnet entries",
 			cidrList: []*net.IPNet{
-				returnIPNetPointers("10.132.0.0/26"),
-				returnIPNetPointers("192.168.0.0/16"),
-				returnIPNetPointers("172.16.0.0/16"),
+				ovntest.MustParseIPNet("10.132.0.0/26"),
+				ovntest.MustParseIPNet("192.168.0.0/16"),
+				ovntest.MustParseIPNet("172.16.0.0/16"),
 			},
 			shouldError: false,
 		},
 		{
 			name: "non-overlapping V6 ipnet entries",
 			cidrList: []*net.IPNet{
-				returnIPNetPointers("fd98:8000::/65"),
-				returnIPNetPointers("fd99::/64"),
-				returnIPNetPointers("fd98:1::/64"),
+				ovntest.MustParseIPNet("fd98:8000::/65"),
+				ovntest.MustParseIPNet("fd99::/64"),
+				ovntest.MustParseIPNet("fd98:1::/64"),
 			},
 			shouldError: false,
 		},
@@ -206,42 +203,42 @@ func Test_checkForOverlap(t *testing.T) {
 		{
 			name: "ipnet entry equal to V4 joinSubnet value",
 			cidrList: []*net.IPNet{
-				returnIPNetPointers("100.64.0.0/16"),
+				ovntest.MustParseIPNet("100.64.0.0/16"),
 			},
 			shouldError: true,
 		},
 		{
 			name: "ipnet entry equal to V6 joinSubnet value",
 			cidrList: []*net.IPNet{
-				returnIPNetPointers("fd98::/64"),
+				ovntest.MustParseIPNet("fd98::/64"),
 			},
 			shouldError: true,
 		},
 		{
 			name: "ipnet entry overlapping with V4 joinSubnet value",
 			cidrList: []*net.IPNet{
-				returnIPNetPointers("100.64.128.0/17"),
+				ovntest.MustParseIPNet("100.64.128.0/17"),
 			},
 			shouldError: true,
 		},
 		{
 			name: "ipnet entry overlapping with V6 joinSubnet value",
 			cidrList: []*net.IPNet{
-				returnIPNetPointers("fd98::/68"),
+				ovntest.MustParseIPNet("fd98::/68"),
 			},
 			shouldError: true,
 		},
 		{
 			name: "ipnet entry encompassing the V4 joinSubnet value",
 			cidrList: []*net.IPNet{
-				returnIPNetPointers("100.0.0.0/8"),
+				ovntest.MustParseIPNet("100.0.0.0/8"),
 			},
 			shouldError: true,
 		},
 		{
 			name: "ipnet entry encompassing the V6 joinSubnet value",
 			cidrList: []*net.IPNet{
-				returnIPNetPointers("fd98::/63"),
+				ovntest.MustParseIPNet("fd98::/63"),
 			},
 			shouldError: true,
 		},
