+++ go-controller/pkg/ovn/allocator/allocator_test.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"net"
 	"testing"
+
+	ovntest "github.com/ovn-org/ovn-kubernetes/go-controller/pkg/testing"
 )
 
 func newSubnetAllocator(clusterCIDR string, hostBits uint32) (*SubnetAllocator, error) {
@@ -34,8 +36,7 @@ func allocateOneNetwork(sna *SubnetAllocator) (*net.IPNet, error) {
 func allocateExpected(sna *SubnetAllocator, n int, expected ...string) error {
 	// Canonicalize expected; eg "fd01:0:0:0::/64" -> "fd01::/64"
 	for i, str := range expected {
-		_, cidr, _ := net.ParseCIDR(str)
-		expected[i] = cidr.String()
+		expected[i] = ovntest.MustParseIPNet(str).String()
 	}
 
 	sns, err := sna.AllocateNetworks()
@@ -328,7 +329,7 @@ func TestMarkAllocatedNetwork(t *testing.T) {
 	}
 
 	// Test subnet that does not belong to network
-	_, subnet, _ := net.ParseCIDR("10.2.3.4/24")
+	subnet := ovntest.MustParseIPNet("10.2.3.0/24")
 	if err := sna.MarkAllocatedNetwork(subnet); err == nil {
 		t.Fatalf("Unexpectedly succeeded in marking allocated subnet that doesn't belong to network (sn=%s)", subnet.String())
 	}
@@ -404,11 +405,11 @@ func TestMultipleSubnets(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	_, sn, _ := net.ParseCIDR("10.1.128.0/18")
+	sn := ovntest.MustParseIPNet("10.1.128.0/18")
 	if err := sna.ReleaseNetwork(sn); err != nil {
 		t.Fatalf("Failed to release the subnet %s: %v", sn.String(), err)
 	}
-	_, sn, _ = net.ParseCIDR("10.2.128.0/18")
+	sn = ovntest.MustParseIPNet("10.2.128.0/18")
 	if err := sna.ReleaseNetwork(sn); err != nil {
 		t.Fatalf("Failed to release the subnet %s: %v", sn.String(), err)
 	}
@@ -460,19 +461,19 @@ func TestDualStack(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	_, sn, _ := net.ParseCIDR("10.1.128.0/18")
+	sn := ovntest.MustParseIPNet("10.1.128.0/18")
 	if err := sna.ReleaseNetwork(sn); err != nil {
 		t.Fatalf("Failed to release the subnet %s: %v", sn.String(), err)
 	}
-	_, sn, _ = net.ParseCIDR("fd01:0:0:3::/64")
+	sn = ovntest.MustParseIPNet("fd01:0:0:3::/64")
 	if err := sna.ReleaseNetwork(sn); err != nil {
 		t.Fatalf("Failed to release the subnet %s: %v", sn.String(), err)
 	}
-	_, sn, _ = net.ParseCIDR("10.2.128.0/18")
+	sn = ovntest.MustParseIPNet("10.2.128.0/18")
 	if err := sna.ReleaseNetwork(sn); err != nil {
 		t.Fatalf("Failed to release the subnet %s: %v", sn.String(), err)
 	}
-	_, sn, _ = net.ParseCIDR("fd01:0:0:7::/64")
+	sn = ovntest.MustParseIPNet("fd01:0:0:7::/64")
 	if err := sna.ReleaseNetwork(sn); err != nil {
 		t.Fatalf("Failed to release the subnet %s: %v", sn.String(), err)
 	}
