+++ go-controller/pkg/ovn/namespace_test.go
@@ -2,7 +2,6 @@ package ovn
 
 import (
 	"fmt"
-	"net"
 
 	"github.com/urfave/cli"
 
@@ -155,8 +154,8 @@ var _ = Describe("OVN Namespace Operations", func() {
 						},
 					},
 				)
-				podMAC, _ := net.ParseMAC("11:22:33:44:55:66")
-				fakeOvn.controller.logicalPortCache.add(tP.nodeName, tP.portName, fakeUUID, podMAC, net.ParseIP(tP.podIP))
+				podMAC := ovntest.MustParseMAC("11:22:33:44:55:66")
+				fakeOvn.controller.logicalPortCache.add(tP.nodeName, tP.portName, fakeUUID, podMAC, ovntest.MustParseIP(tP.podIP))
 				fakeOvn.controller.WatchNamespaces()
 
 				_, err := fakeOvn.fakeClient.CoreV1().Namespaces().Get(namespaceT.Name, metav1.GetOptions{})
